pyproject.toml
[project]
name = "carrinho-do-dudu-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kaubarros-24", email = "martinsbarroskaua@gmail.com"},
]
dependencies = ["fastapi>=0.115.12", "uvicorn[standard]>=0.34.2", "gunicorn>=23.0.0", "sqlalchemy>=2.0.40", "passlib[bcrypt]>=1.7.4", "python-jose[cryptography]>=3.4.0", "pydantic>=2.11.3"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
